♻✅ Refactor tests in various files
