+++ som_leads_polissa/tests/__init__.py
@@ -1,3 +1,9 @@
 # -*- coding: utf-8 -*-
 from __future__ import absolute_import
-from .tests_som_lead_www import *
+
+from .test_lead_www_creation import *
+from .test_lead_www_self_consumption import *
+from .test_lead_www_member_linking import *
+from .test_lead_www_member_payment import *
+from .test_lead_www_validation import *
+from .test_lead_www_misc import *
